Break up cohort page class
