+++ util/spec_helper.rb
@@ -59,6 +59,7 @@ require_relative '../pages/junction/canvas_e_grades_export_page'
 require_relative '../pages/boac/boac_pages'
 require_relative '../pages/boac/home_page'
 require_relative '../pages/boac/cohort_page'
+require_relative '../pages/boac/cohort_list_view_page'
 require_relative '../pages/boac/cohort_matrix_page'
 require_relative '../pages/boac/student_page'
 require_relative '../pages/boac/api_user_analytics_page'
